requirements.txt
Flask==2.2.5
numpy==1.24.2
Pillow==9.3.0
tensorflow==2.10.0
Werkzeug==2.2.5
